SuperLearner wrapper for antibody acquisition models

An internal SuperLearner wrapper for the antibody acquisition and loss model in Yman et al. 2016

SL.Yman2016(Y, X, newX = NULL, ...)

Arguments

Y
log quantitative antibody level

X
age

newX
(optional) new values of age over which to predict

...
other arguments passed to the function (currently ignored)

Value

pred Predicted outcome values fit Maximum likelihood fit object returned from optim

Details

The SL.Yman2016 wrapper enables the inclusion of this model in a SuperLearner ensemble. The model assumes that rates of antibody acquistion and loss are constant. It can only use information about age and not additional covariates. If X includes more than 1 column, then the function will assume that the first column of X is age and will throw a warning.

See also

Yman2016 SuperLearner