dataset: Garki Project (Nigeria) serology data

The garki_sero dataset includes a semi-processed version of publicly available serology data from the Garki Project, Nigeria. This file includes merged data from the village, people, and serology tables in the original Garki files. The original files are documented through the Garki Project's website (http://garkiproject.nd.edu/). These data were downloaded originally from the Swiss Tropical Institute (http://www.swisstph.ch/fr/ressources/epidemiological-databases.html). The dictionary file from STI included important additional information about data labels. The Stata script that merged the original tables is located here: https://osf.io/7frw2/.

data(garki_sero)

Format

A data frame with 20,441 observations and 52 variables. Variable labels below have been copied from http://garkiproject.nd.edu. The original table's source is added as a prefix, else it was derived.

id

people: individual ID

village

village: village ID

compount

people: compound ID

serosvy

serology: Serological survey number

survey

village: Garki project survey number

area

village: Garki project study area

grid

village: Garki project grid number

fup

village: The follow-up cluster village is in

vname

village: Village name

vu

village: Code representing one village unit for a group

grouped1

village: number of compounds in compact part? CD

grouped2

village: unknown

scattered1

village: number of compounds in scattered part? CD

scattered2

village: unknown

population

village: Population of village (when?) CD

move

people: unknown

pers

people: unknown

sfreg

people: The number of the survey the person first registered with.

sex

people: Identifies the sex of the person. Coding unknown.

dob

people: unknown

dateb

people: The date of birth of the person.

batch_igg

serology: Batch number IgG.

batch_igm

serology: Batch number IgM.

plate_igg

serology: Plate number IgG.

plate_igm

serology: Plate number IgM.

diam_igg

serology: Specimen diameter IgG.

diam_igm

serology: Specimen diameter IgM.

eval_igg

serology: Specimen evaluation IgG.

eval_igm

serology: Specimen evaluation IgM.

perc_igg

serology: Percentage IgG.

perc_igm

serology: Percentage IgM.

precipitin

serology: Result of the precipitin reaction. Coded as follows: 0=Test not done, 1-7=Number of bands, 8=No bands

blank1

serology: unknown

ifat_pfa

serology: IFA test for Plasmodium falciparum, 0=no result, 1= Titre 20, 2=60, 3=180, 4=540, 5=1620,6=4860, 7=14580,8=titres >14580, 9= titre < 20 (for computations take 0); In general Titre=20*3**(code-1)

blank2

serology: unknown

ifat_pm

serology: IFA test for Plasmodium malariae, 0=no result, 1= Titre 20, 2=60, 3=180, 4=540, 5=1620,6=4860, 7=14580,8=titres >14580, 9= titre < 20 (for computations take 0); In general Titre=20*3**(code-1)

haem_code

serology: 1 = AA; 2 = AS, 3 = SS, 4 =AC, 5 = CC, 6 = SC, 9 = not done

haem_phbs

serology: unknown

recd_iha

serology: Record identifier for IHA (code 116)

specno

serology: specimen number

titrcode

serology: IHA titre code, 4 to 16 = titre 2**4 to 2**16 (in general titre=2**code)

date

serology: unknown

tr

(derived): factor: control or intervention village

phase

(derived): factor: baseline, intervention, post-intervention phase

wetseason

(derived): factor: dry or wet season

bdate

(derived): individuals date of birth (standard date format)

mdate

(derived): measurement date

agedays

(derived): age in days at time of measurement

ageyrs

(derived): age in years at time of measurement

ifatpftitre

(derived): IFA test for P. falciparum antibody titres

ifatpmtitre

(derived): IFA test for P. malariae antibody titres

ihatitre

(derived): IHA test antibody titres

Source

http://www.swisstph.ch/fr/ressources/epidemiological-databases.html

References

Molineaux L, Gramiccia G, 1980. The Garki Project. Geneva: World Health Organisation. http://whqlibdoc.who.int/publications/9241560614.pdf

http://garkiproject.nd.edu/