dataset: Mauke island lymphatic filariasis Wb123 antibody measurements

Serological measurements from two surveys on Mauke, Cook Islands The dataset includes serum samples from two cross-sectional measurements of the permanent resident population; the first in 1975 (N=362, approximately 58

data(mauke_wb123)

Format

A data frame with 922 observations and 7 variables.

id75

individual ID in 1975

id92

individual ID in 1992

year

Year of the measurement (1975 or 1992)

age

age in years. Note that age records years completed.

CAg

Level of circulating filarial antigen (>32 is considered positive for circulating antigen)

wb123

response to the Wuchereria bancrofti Wb123 antigen

wb123pos

Seropositive, based on Wb123 using a cutoff of 10968, identified in Kubofcik et al. (2012)

References

C. Steel, J. Kubofcik, E. A. Ottesen, T. B. Nutman, Antibody to the filarial antigen Wb123 reflects reduced transmission and decreased exposure in children born following single mass drug administration (MDA). PLoS Negl. Trop. Dis. 6, e1940 (2012).

J. Kubofcik, D. L. Fink, T. B. Nutman, Identification of Wb123 as an early and specific marker of Wuchereria bancrofti infection. PLoS Negl. Trop. Dis. 6, e1930 (2012).